+++ custom_components/recalbox/api.py
@@ -14,6 +14,7 @@ class RecalboxAPI:
                  udp_recalbox: int = 1337, # https://github.com/recalbox/recalbox-api
                  udp_retroarch: int = 55355, # https://docs.libretro.com/development/retroarch/network-control-interface/
                  api_port_kodi: int = 8081, # https://kodi.wiki/view/JSON-RPC_API
+                 only_ip_v4: bool = False,
                  ):
         self.host = host
         self.api_port_os = api_port_os # Arrêter, Reboot de Recalbox...
@@ -21,6 +22,14 @@ class RecalboxAPI:
         self.udp_recalbox = udp_recalbox # Lancer une ROM
         self.udp_retroarch = udp_retroarch
         self.api_port_kodi = api_port_kodi # Pour quitter Kodi
+        self.only_ip_v4 = only_ip_v4
+
+
+    def _getSocketType(self):
+        if self.only_ip_v4:
+            return socket.AF_INET
+        else :
+            return socket.AF_UNSPEC
 
 
     async def send_udp_command(self, port, message):
@@ -29,12 +38,12 @@ class RecalboxAPI:
         transport, protocol = await loop.create_datagram_endpoint(
             lambda: asyncio.DatagramProtocol(),
             remote_addr=(self.host, port),
-            family=socket.AF_INET  # Force la résolution en IPv4
+            family=self._getSocketType()
         )
         try:
             transport.sendto(message.encode())
             return True
-        except:
+        except Exception as e:
             _LOGGER.error(f"Fail to send UDP message to {self.host} on port {port}")
             return False
         finally:
@@ -44,12 +53,12 @@ class RecalboxAPI:
     async def post_api(self, path, port=80):
         url = f"http://{self.host}:{port}{path}"
         _LOGGER.debug(f"API POST {url}")
-        connector = aiohttp.TCPConnector(family=socket.AF_INET) # Force la résolution en IPv4
+        connector = aiohttp.TCPConnector(family=self._getSocketType()) # Force la résolution en IPv4
         async with aiohttp.ClientSession(connector=connector) as session:
             try:
                 async with session.post(url) as response:
                     return response.status == 200
-            except:
+            except Exception as e:
                 _LOGGER.error(f"Failed to call {url}")
                 raise
 
@@ -57,15 +66,15 @@ class RecalboxAPI:
     async def get_roms(self, console):
         url = f"http://{self.host}:{self.api_port_gamesmanager}/api/systems/{console}/roms"
         _LOGGER.debug(f"API GET roms from {url}")
-        connector = aiohttp.TCPConnector(family=socket.AF_INET) # Force la résolution en IPv4
+        connector = aiohttp.TCPConnector(family=self._getSocketType()) # Force la résolution en IPv4
         async with aiohttp.ClientSession(connector=connector) as session:
             try:
                 async with session.get(url, timeout=10) as response:
                     if response.status == 200:
                         data = await response.json()
                         return data.get("roms", [])
-            except:
-                _LOGGER.error(f"Failed to get roms list on {url}")
+            except Exception as e:
+                _LOGGER.error(f"Failed to get roms list on {url} : {e}")
                 raise
 
 
@@ -77,15 +86,15 @@ class RecalboxAPI:
             "id": 1
         }
         _LOGGER.debug(f"API to quit Kodi : {kodi_url}")
-        connector = aiohttp.TCPConnector(family=socket.AF_INET) # Force la résolution en IPv4
+        connector = aiohttp.TCPConnector(family=self._getSocketType()) # Force la résolution en IPv4
         async with aiohttp.ClientSession(connector=connector) as session:
             try:
                 async with session.post(kodi_url, json=payload, timeout=5) as response:
                     if response.status == 200:
                         await asyncio.sleep(5)
                         return True
-            except:
-                _LOGGER.error(f"Failed to quit Kodi via JSON RPC on {kodi_url}")
+            except Exception as e:
+                _LOGGER.error(f"Failed to quit Kodi via JSON RPC on {kodi_url} : {e}")
                 return False
 
 
@@ -97,14 +106,14 @@ class RecalboxAPI:
             "id": 1
         }
         _LOGGER.debug(f"Ping Kodi : {kodi_url}")
-        connector = aiohttp.TCPConnector(family=socket.AF_INET) # Force la résolution en IPv4
+        connector = aiohttp.TCPConnector(family=self._getSocketType()) # Force la résolution en IPv4
         async with aiohttp.ClientSession(connector=connector) as session:
             try:
                 async with session.post(kodi_url, json=payload, timeout=5) as response:
                     if response.status == 200:
                         return True
-            except:
-                _LOGGER.info(f"Failed to ping Kodi via JSON RPC on {kodi_url}")
+            except Exception as e:
+                _LOGGER.info(f"Failed to ping Kodi via JSON RPC on {kodi_url} : {e}")
                 return False
 
 
@@ -113,7 +122,7 @@ class RecalboxAPI:
     async def get_current_status(self):
         url = f"http://{self.host}:{self.api_port_gamesmanager}/api/status"
         _LOGGER.debug(f"API GET current Recalbox status {url}")
-        connector = aiohttp.TCPConnector(family=socket.AF_INET) # Force la résolution en IPv4
+        connector = aiohttp.TCPConnector(family=self._getSocketType()) # Force la résolution en IPv4
         # {
         #   "Action": "rungame",
         #   "Parameter": "/recalbox/share/roms/megadrive/001 Sonic 1.bin",
@@ -166,8 +175,8 @@ class RecalboxAPI:
                             "scriptVersion": None,
                             "status": "ON"
                         }
-            except:
-                _LOGGER.error(f"Failed to get recalbox status on API {url}")
+            except Exception as e:
+                _LOGGER.error(f"Failed to get recalbox status on API {url} ({e})")
                 if (await self.is_kodi_running()) :
                     _LOGGER.debug(f"Kodi seems to be running ! Simulating JSON data for Recalbox HA status")
                     return {
@@ -190,7 +199,7 @@ class RecalboxAPI:
     async def ping(self) -> bool:
         """Exécute un ping système vers l'hôte."""
         _LOGGER.debug(f"PING recalbox on {self.host}")
-        command = f"ping -4 -c 1 -W 1 {self.host} > /dev/null 2>&1"
+        command = f"ping {'-4 ' if self.only_ip_v4 else ''}-c 1 -W 1 {self.host} > /dev/null 2>&1"
         try:
             # On exécute la commande système de façon asynchrone
             process = await asyncio.create_subprocess_shell(command)
@@ -212,7 +221,7 @@ class RecalboxAPI:
             for port in TCP_PORTS:
                 try:
                     _LOGGER.debug(f"Testing TCP port {port} on {self.host}")
-                    conn = asyncio.open_connection(self.host, port, family=socket.AF_INET)
+                    conn = asyncio.open_connection(self.host, port, family=self._getSocketType())
                     _reader, writer = await asyncio.wait_for(conn, timeout=1.0)
                     writer.close()
                     await writer.wait_closed()
@@ -229,6 +238,6 @@ class RecalboxAPI:
                     return False
 
             return True
-        except:
-            _LOGGER.debug(f"Failed to PING ports of {self.host}")
+        except Exception as ex:
+            _LOGGER.debug(f"Failed to PING ports of {self.host} : {ex}")
             return False